tests/requirements.txt
# Test dependencies for Insurance DW Pipeline integration tests
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-timeout>=2.1.0
psycopg2-binary>=2.9.0
